% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{bacmeta_path_object}
\alias{bacmeta_path_object}
\title{Return the bacmeta environment}
\usage{
bacmeta_path_object()
}
\value{
The environment that contains the path to compiled bacmeta
}
\description{
Return the bacmeta environment
}
