% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{compile_bacmeta}
\alias{compile_bacmeta}
\title{Compile bacmeta}
\usage{
compile_bacmeta(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{TRUE}, prints no compilation progress to console.
Default is \code{FALSE}.}
}
\description{
Copies bacmeta source files to an environment-specific temporary directory
and compiles bacmeta there. The path to the compilation directory is saved
in the \code{.sim_env} object and is accessible via
\code{path_to_bacmeta()}. If bacmeta is already compiled, does nothing.
}
