% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{compile_isource}
\alias{compile_isource}
\title{Compile isource}
\usage{
compile_isource(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{TRUE}, prints no compilation progress to console.
Default is \code{FALSE}.}
}
\description{
Copies isource source files to an environment-specific temporary directory
and compiles isource there. The path to the compilation directory is saved
in the \code{.bacmeta_env} object and is accessible via
\code{path_to_isource()}. If isource is already compiled, does nothing.
}
