% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{create_migration_input}
\alias{create_migration_input}
\title{Generate a Migration Rate Matrix File for Running Bacmeta Simulation}
\usage{
create_migration_input(rates = NULL, out_path = getwd(), suffix = NULL)
}
\arguments{
\item{rates}{the migration rates to associate with each population
pair.  If \code{NULL} (default), the resulting matrix will
contain only zeros. Otherwise, a square matrix is expected.}

\item{out_path}{The destination to which the migration file will
be written. Must be an existing directory.}

\item{suffix}{A suffix that will be used for the migration file
name. If \code{suffix} is \code{NULL} (default), the filename
will simply be "migration.input", otherwise the suffix will be
pasted in between "migration" and ".input". E.g. if
\code{suffix} is 123, the filename will be
"migration123.input". All symbols in \code{suffix} must be
alphanumeric (A-Z, a-z, 0-9).}
}
\value{
\code{out_path}
}
\description{
Creates a square matrix of dimensions \code{n_populations} x
\code{n_populations} where each element represents a migration
rate as defined in the \code{rates} parameter. The rate at row i,
column j represents migration from population i to population
j. The matrix is written to a tab-separated file named
migration[\code{suffix}].input in the directory \code{out_path}.
}
\details{
This migration file can then be used in a Bacmeta simulation with
\code{n_populations} populations, in place of a single migration
rate.  See Bacmeta readme for details.
}
\author{
Wiktor Gustafsson
}
