% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceSim.R, R/utils.R
\name{get_sequences}
\alias{get_sequences}
\title{Extract Sequences from FASTA Files}
\usage{
get_sequences(
  path = system.file("extdata/MLST_alleles", package = "sourceSim")
)

get_sequences(
  path = system.file("extdata/MLST_alleles", package = "sourceSim")
)
}
\arguments{
\item{path}{the path to the directory containing the FASTA files}
}
\value{
a named character vector where each element is a concatenation of
all sequences found in each file in \code{path}, i.e., every non-
header line of the file. The name of each element is the file name.

a named character vector where each element is a concatenation of
all sequences found in each file in \code{path}, i.e., every non-
header line of the file. The name of each element is the file name.
}
\description{
Extract all sequence data from all FASTA files in a given directory.

Extract all sequence data from all FASTA files in a given directory.
}
\author{
Wiktor Gustafsson

Wiktor Gustafsson
}
