% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hald.R
\name{hald.list}
\alias{hald.list}
\title{run hald model on a \code{list} object using OpenBUGS}
\usage{
\method{hald}{list}(
  x,
  iter = 10000,
  burnin = iter/2,
  thinning = 10,
  n_chains = 3,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{x}{the \code{list} object}

\item{iter}{the number of iterations to run the simulation for}

\item{burnin}{the number of burnin iterations}

\item{thinning}{the thinning rate}

\item{n_chains}{the number of simulation chains}

\item{simplify}{When TRUE the return is a matrix of
sumamry stats. When FALSE the return is a list with the same
matrix and the simulation results.}

\item{...}{other arguments}
}
\value{
proportions of the attribution for each population
}
\description{
run hald model on a \code{list} object using OpenBUGS
}
