% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hald.R
\name{hald}
\alias{hald}
\title{hald functions}
\usage{
hald(x, ...)
}
\arguments{
\item{x}{An object to run the hald method on}

\item{...}{other arguments}
}
\value{
A character vector of the proportions
}
\description{
hald functions
}
