% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hald.R
\name{hald.sourceSim_result}
\alias{hald.sourceSim_result}
\title{run hald model on a \code{sourceSim_result} object}
\usage{
\method{hald}{sourceSim_result}(
  x,
  iter = 10000,
  burnin = iter/2,
  thinning = 10,
  n_chains = 3,
  q_max = 2000,
  a_max = 100,
  others_cutoff = 4,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{x}{the \code{sourceSim_result} object}

\item{iter}{the number of iterations to run the simulation for}

\item{burnin}{the number of burnin iterations}

\item{thinning}{the thinning rate}

\item{n_chains}{the number of simulation chains}

\item{others_cutoff}{For MLST's that have less than or equal to
\code{others_cutoff} individuals, reclassify them as type "others"}

\item{simplify}{When TRUE the return is a matrix of
summary stats. When FALSE the return is a list with the same
matrix and the mcmc results.}

\item{...}{other arguments}

\item{q_MaxRange}{the maximum value of the Hald model 'q' parameter
source-specific coefficient)}

\item{a_MaxRange}{the maximum value of the Hald model 'a' parameter
(MLST-specific coefficient)}
}
\value{
proportions of the attribution for each population
}
\description{
run hald model on a \code{sourceSim_result} object
}
