% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{internal_sequences}
\alias{internal_sequences}
\title{internal_sequences}
\usage{
internal_sequences(species)
}
\arguments{
\item{species}{The species whose sequences you want to extract, which must
be the name of the folder in which the sequences are contained.
Currently the only option is \code{"Campylobacter"}.}
}
\value{
a named character with all the sequences (see
\code{get_sequences()} documentation for details).
}
\description{
Extract SourceSim-internal MLST sequences pertaining to a specific built-in
species.
}
