% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{is_bacmeta_compiled}
\alias{is_bacmeta_compiled}
\title{Check if Bacmeta is Compiled}
\usage{
is_bacmeta_compiled()
}
\value{
\code{TRUE} if bacmeta is compiled, otherwise \code{FALSE}.
}
\description{
If bacmeta is compiled, each C++ script in its src directory should have
a corresponding .o file, and there should be a "Simu" file in bacmeta root.
}
