% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{is_isource_compiled}
\alias{is_isource_compiled}
\title{Check if isource is Compiled}
\usage{
is_isource_compiled()
}
\value{
\code{TRUE} if bacmeta is compiled, otherwise \code{FALSE}.
}
\description{
If isource is compiled, each C++ script in its src directory should have
a corresponding .o file, and there should be a "isource" file in isource root.
}
