% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{isource.data.frame}
\alias{isource.data.frame}
\title{isource.data.frame}
\usage{
\method{isource}{data.frame}(
  x = NULL,
  pops,
  iter = 10000,
  burnin = iter/2,
  thinning = 50,
  dirichlet_param = 1,
  group_var = "group",
  simplify = TRUE,
  ...
)
}
\arguments{
\item{x}{The result of a simulation of data, a \code{data.frame}}

\item{pops}{the names of the populations}

\item{iter}{The number of iterations to run}

\item{burnin}{The burin length}

\item{thinning}{The thinning rate}

\item{dirichlet_param}{The parameter on the dirichlet}

\item{group_var}{The variable to group the results by}

\item{simplify}{When TRUE the return is a matrix of
summary stats. When FALSE the return is a list with the same
matrix and the mcmc results.}

\item{...}{other arguments}
}
\value{
proportions of the attribution for each population
}
\description{
Runs isource asymmetric island model on a dataframe
}
