% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{isource}
\alias{isource}
\title{isource}
\usage{
isource(
  x = NULL,
  iter = 20000,
  burnin = 1000,
  thinning = 50,
  dirichlet_param = 1,
  group_var = "group"
)
}
\arguments{
\item{x}{a sourceSim_result object}

\item{iter}{the number of iterations}

\item{burnin}{the burnin}

\item{thinning}{the thinning rate}

\item{dirichlet_param}{the parameter to the dirichlet}

\item{group_var}{the name of the grouping variable}
}
\description{
Runs isource asymmetric island model:
}
