% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{isource.sourceSim_result}
\alias{isource.sourceSim_result}
\title{isource.sourceSim_result}
\usage{
\method{isource}{sourceSim_result}(
  x = NULL,
  iter = 20000,
  burnin = 1000,
  thinning = 50,
  dirichlet_param = 1,
  group_var = "group"
)
}
\arguments{
\item{x}{The result of a simulation of data}

\item{iter}{The number of iterations to run}

\item{burnin}{The burin length}

\item{thinning}{The thinning rate}

\item{dirichlet_param}{The parameter on the dirichlet}

\item{group_var}{The variable to group the results by}
}
\description{
Runs isource asymmetric island model:
}
