% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.sourceSim_result}
\alias{plot.sourceSim_result}
\title{plot}
\usage{
\method{plot}{sourceSim_result}(x, piecol = NULL, type = "fan", pie_cex = 1, legend = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.frame} with bacmeta simulation result sequences,
as produced by \code{read_results()}.}

\item{piecol}{A vector of colours}

\item{type}{"fan" (the default) or other options from plot.phylo in ape}

\item{legend}{A logical if a legend should be included}

\item{...}{Other arguments}

\item{pie.cex}{a scale parameter for the size of the pies}
}
\value{
A "\code{recordedplot}" object with the phylogenetic tree.
}
\description{
Plots a phylogenetic tree of sequences that are the result of a bacmeta
simulation
}
