% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_results}
\alias{plot_results}
\title{plot_results}
\usage{
plot_results(df)
}
\arguments{
\item{df}{A \code{data.frame} with bacmeta simulation result sequences,
as produced by \code{read_results()}.}
}
\value{
A "\code{recordedplot}" object with the phylogenetic tree.
}
\description{
Plots a phylogenetic tree of sequences that are the result of a bacmeta
simulation
}
