% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.sourceSim_result}
\alias{print.sourceSim_result}
\title{print.sourceSim_result}
\usage{
\method{print}{sourceSim_result}(x, ...)
}
\arguments{
\item{x}{A bacmeta simulation result}

\item{...}{Other arguments}
}
\description{
print.sourceSim_result
}
