% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{rdirichlet}
\alias{rdirichlet}
\title{rdirichlet}
\usage{
rdirichlet(n = 1, alpha = c(1, 1))
}
\arguments{
\item{n}{The number of samples from the dirichlet}

\item{alpha}{The shape of each gamma distribution. The length
dictates the number of bins in the dirichlet.}
}
\value{
A matrix
}
\description{
sample from a dirichlet distribution. Returns a matrix where each
row is a sample. The definition of a dirichlet is that each set of
n samples must sum to 1. Therefore the rowSums(rdirichlet()) is
always 1.
}
