% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{read_paramfile}
\alias{read_paramfile}
\title{Read (custom or default) bacmeta simulation parameter file.}
\usage{
read_paramfile(path = "default", as_list = FALSE)
}
\arguments{
\item{path}{Either the word "default" (which is the, well, default),
which fetches the "default.params" file found in the bacmeta
directory which is bundled with this package. OR a path to a valid
parameter file.}

\item{as_list}{Return the results as a list? If \code{FALSE} (default),
simply returns the parameters as a two-column data.frame, where
column 1 is the parameter names and column 2 is their values. If
\code{TRUE}, instead returns a name list where each element
is a parameter.}
}
\description{
Read (custom or default) bacmeta simulation parameter file.
}
