% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{read_results}
\alias{read_results}
\title{read_results}
\usage{
read_results(sequences, initial)
}
\arguments{
\item{sequences}{A path to a "Sequences" file which has been output from
running bacmeta.}

\item{initial}{A path to an "InitialSequences" file which has been output
from running bacmeta. This is only used to match against the data
in \code{sequences} and exclude the initial sequences from the
dataset.}
}
\value{
a \code{data.frame} with the information described above.
}
\description{
Reads the sequence output of bacmeta simulation and produces a summary
dataset. Each row in the dataset represents a genotype produced in the
simulation. One genotype is one unique combination of locus alleles. The
dataset contains information about:
\itemize{
\item{Pop_n} The number of occurences of the genotype in each simulated
population, in columns named "Pop_0", "Pop_1", etc;
\item{total} the total number of occurences (sum of all population columns);
\item{seqID} A unique numeric genotype/sequence ID;
\item{MLST} An "MLST" ID which is a concatenation of the unique allele IDs of
all simulated loci, separated by hyphens;
\item{seq} A DNA sequence which is simply the sequences of all alleles
pasted together.
}
\code{seqID}, \code{MLST} and \code{seq} are all unique identifiers.
}
