% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\docType{data}
\name{result}
\alias{result}
\title{Output data from a simulation}
\source{
From a simulation
}
\description{
The output of a simulation of 20000 generation of \emph{C. jejuni} from
3 metapopulations with variable overlap
}
\keyword{datasets}
