% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(ob, exp)
}
\arguments{
\item{ob}{a numeric vector of observations}

\item{exp}{a numeric expected value. Must be of length 1.}
}
\value{
a numeric value
}
\description{
Calculate the root mean square error of a series of observations to an
expected value. The formula is that of standard deviation, but variation
is measured around an expected (given) value rather than the mean of the
data.
}
