% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_humans.R
\name{sample_humans}
\alias{sample_humans}
\title{functions to sample humans}
\usage{
sample_humans(x, ...)
}
\arguments{
\item{x}{An object to sample humans on}

\item{...}{other arguments}
}
\value{
A character vector of the proportions
}
\description{
functions to sample humans
}
