% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_humans.R
\name{sample_humans.sourceSim_result}
\alias{sample_humans.sourceSim_result}
\title{sample_humans}
\usage{
\method{sample_humans}{sourceSim_result}(x, attribution = 1, n = 100, overwrite = FALSE, ...)
}
\arguments{
\item{x}{A result of a \code{simu()} function, a \code{sourceSim_result}
object}

\item{attribution}{the attribution fraction for each of the source
populations}

\item{n}{The number of human cases}

\item{overwrite}{If humans have already been sampled for this simulation,
should a new sampling be run? Will overwrite the old one. Default
is \code{FALSE}, in which case nothing happens.}
}
\value{
A sourceSim_result object with added humans
}
\description{
Sample the human cases from the bacterial sources given an
attribution fraction.
}
