% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  input,
  out_path,
  migration = NA_character_,
  simu_dir = tempdir(),
  keep_simufiles = FALSE,
  plot = FALSE
)
}
\description{
simulate
}
