% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{bacmeta_path_object}
\alias{bacmeta_path_object}
\title{Return the bactmeta environment}
\usage{
bacmeta_path_object()
}
\value{
The evironment that contains the path to compiled bactmeta
}
\description{
Return the bactmeta environment
}
