% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hald.R
\name{hald.list}
\alias{hald.list}
\title{run hald model on a \code{list} object}
\usage{
\method{hald}{list}(x, iter = 20000, burnin = 1000, thinning = 50, n_chains = 3, ...)
}
\arguments{
\item{x}{the \code{list} object}

\item{iter}{the number of iterations to run the simulation for}

\item{burnin}{the number of burnin iterations}

\item{thinning}{the thinning rate}

\item{n_chains}{the number of simulation chains}

\item{...}{other arguments}
}
\value{
a \code{hald_output} object, the result of an OpenBUGS run with
the Hald model.
}
\description{
run hald model on a \code{list} object
}
