% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hald.R
\name{hald.sourceSim_result}
\alias{hald.sourceSim_result}
\title{run hald model on a \code{sourceSim_result} object}
\usage{
\method{hald}{sourceSim_result}(
  x,
  iter = 20000,
  burnin = 1000,
  thinning = 50,
  n_chains = 3,
  q_MaxRange = 2000,
  a_MaxRange = 100,
  others_cutoff = 4,
  ...
)
}
\arguments{
\item{x}{the \code{sourceSim_result} object}

\item{iter}{the number of iterations to run the simulation for}

\item{burnin}{the number of burnin iterations}

\item{thinning}{the thinning rate}

\item{n_chains}{the number of simulation chains}

\item{q_MaxRange}{the maximum value of the Hald model 'q' parameter
source-specific coefficient)}

\item{a_MaxRange}{the maximum value of the Hald model 'a' parameter
(MLST-specific coefficient)}

\item{others_cutoff}{For MLST's that have less than or equal to
\code{others_cutoff} individuals, reclassify them as type "others"}

\item{...}{other arguments}
}
\value{
a \code{hald_output} object, the result of an OpenBUGS run with
the Hald model.
}
\description{
run hald model on a \code{sourceSim_result} object
}
