% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{isource}
\alias{isource}
\title{isource functions}
\usage{
isource(x, ...)
}
\arguments{
\item{x}{An object to run the isource method on}

\item{...}{other arguments}
}
\value{
A character vector of the proportions
}
\description{
isource functions
}
