% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{plot.isource_output}
\alias{plot.isource_output}
\title{plot.isource_output}
\usage{
\method{plot}{isource_output}(
  x,
  type = c("mcmc", "hist", "bar", "evol", "pie-evol", "area"),
  COL = NULL,
  cod = NULL,
  ...
)
}
\arguments{
\item{x}{The result of the isource method}

\item{type}{One of c("mcmc", "hist", "bar", "evol", "pie-evol", "area")}

\item{COL}{A colour vector}

\item{cod}{The order of the groups}

\item{...}{other arguments}
}
\description{
Plots described in the isource documentation
}
