% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.sourceSim_result}
\alias{plot.sourceSim_result}
\title{plot}
\usage{
\method{plot}{sourceSim_result}(x, piecol = NULL, legend = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.frame} with bacmeta simulation result sequences,
as produced by \code{read_results()}.}

\item{...}{Other arguments}
}
\value{
A "\code{recordedplot}" object with the phylogenetic tree.
}
\description{
Plots a phylogenetic tree of sequences that are the result of a bacmeta
simulation
}
