% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isource.R
\name{sample_humans}
\alias{sample_humans}
\title{sample_humans}
\usage{
sample_humans(x, attribution = 1, n = 100)
}
\arguments{
\item{x}{A result of a simu() function}

\item{attribution}{the attribution fraction for each of the source
populations}

\item{n}{The number of human cases}
}
\value{
A sourceSim_result object with added humans
}
\description{
Sample the human cases from the bacterial sources given an
attribution fraction.
}
