% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacmeta.R
\name{simu}
\alias{simu}
\title{simu}
\usage{
simu(input = NULL, migration = NULL)
}
\arguments{
\item{input}{The parameters for running bacmeta.  Accepts one of
two options:
\itemize{
\item \code{NULL} (default), which uses the
\code{default.params} which is included in the package;
\item a named list of parameters, where each name is a
bacmeta-accepted parameter, and the values are numeric
(note: parameters affecting file outputs from the
simulation, i.e. \code{SEQI} and \code{ISEQ}, will be
ignored to ensure the desired results can be captured).
}}

\item{migration}{Migration matrix for bacmeta simulation. Only
used if the \code{MIGI} parameter in the main bacmeta
parameter file is set to 1, otherwise ignored. Accepts one of
two options:
\itemize{
\item \code{NULL} (default), which creates a zero matrix
meaning there is no migration between any populations. This
requires the 'NPOP' parameter to be set in the main parameter file.
\item a numeric matrix of dimensions n x n,
where n is the number of populations (as defined by the 'NPOP'
parameter in the main parameter file).
}}
}
\value{
A \code{data.frame} with the resulting DNA sequences from
running the simulation (see \code{read_results() function for
    details}).
}
\description{
Runs bacmeta simulation:
\enumerate{
\item Verifies that the supplied parameter files are in a valid format,
\item copies the supplied bacmeta parameter files to a simulation
directory,
\item compiles bacmeta if necessary,
\item runs bacmeta,
\item copies outputs from the simulation to a specified output directory.
}
}
